How to Perform Object Detection With YOLOv3 in Keras
Object detection is a task in computer vision that involves identifying the presence, location, and type of one or more objects in a given photograph.

It is a challenging problem that involves building upon methods for object recognition (e.g. where are they), object localization (e.g. what are their extent), and object classification (e.g. what are they).

In recent years, deep learning techniques are achieving state-of-the-art results for object detection, such as on standard benchmark datasets and in computer vision competitions. Notable is the “You Only Look Once,” or YOLO, family of Convolutional Neural Networks that achieve near state-of-the-art results with a single end-to-end model that can perform object detection in real-time.

In this tutorial, you will discover how to develop a YOLOv3 model for object detection on new photographs.

After completing this tutorial, you will know:

YOLO-based Convolutional Neural Network family of models for object detection and the most recent variation called YOLOv3.
The best-of-breed open source library implementation of the YOLOv3 for the Keras deep learning library.
How to use a pre-trained YOLOv3 to perform object localization and detection on new photographs.

https://machinelearningmastery.com/how-to-perform-object-detection-with-yolov3-in-keras/